requirements.txt
# gRPC and Protocol Buffers
grpcio==1.59.0
grpcio-tools==1.59.0
grpcio-reflection==1.59.0
protobuf==4.24.4

# Database
sqlalchemy==2.0.23
# sqlite3 is built into Python

# Password hashing (equivalent to bcrypt in Go)
bcrypt==4.1.1

# RabbitMQ client
pika==1.3.2

# Type hints
typing-extensions==4.8.0